COTECH-667 make JWP strategy rules via Optimizely detection sooner
