+++ src/platforms/shared/index.ts
@@ -20,6 +20,7 @@ export * from './setup/ad-engine-stack.setup';
 export * from './setup/base-context.setup';
 export * from './setup/bidders-targeting-updater.service';
 export * from './setup/gpt.setup';
+export * from './setup/jwp-strategy-rules.setup';
 export * from './setup/labrador.setup';
 export * from './setup/load-times.setup';
 export * from './setup/metric-reporter.setup';
