+++ src/platforms/shared/setup/player.setup.ts
@@ -10,7 +10,6 @@ import {
 	InstantConfigService,
 	JWPlayerManager,
 	jwpSetup,
-	Optimizely,
 	slotService,
 	utils,
 	VastResponseData,
@@ -21,23 +20,11 @@ import { Injectable } from '@wikia/dependency-injection';
 
 const logGroup = 'player-setup';
 
-const OPTIMIZELY_STRATEGY_RULES_EXPERIMENT = {
-	EXPERIMENT_ENABLED: 'strategy_rules',
-	EXPERIMENT_VARIANT: 'strategy_rules_variant',
-};
-
-const OPTIMIZELY_STRATEGY_RULES_EXPERIMENT_VARIANTS = {
-	DISABLED: 'strategy_rules_disabled',
-	ENABLED: 'strategy_rules_enabled',
-	UNDEFINED: 'strategy_rules_undefined',
-};
-
 @Injectable()
 export class PlayerSetup extends BaseServiceSetup {
 	constructor(
 		protected instantConfig: InstantConfigService,
 		protected globalTimeout: utils.GlobalTimeout,
-		protected optimizely: Optimizely,
 		protected jwpManager: JWPlayerManager,
 		protected vastTaglessRequest: VastTaglessRequest,
 	) {
@@ -46,8 +33,6 @@ export class PlayerSetup extends BaseServiceSetup {
 	}
 
 	async call() {
-		this.setupOptimizelyExperiment();
-
 		const showAds = !context.get('options.wad.blocking');
 		const vastResponse: VastResponseData =
 			showAds &&
@@ -152,24 +137,5 @@ export class PlayerSetup extends BaseServiceSetup {
 			targetingParams: utils.getCustomParameters(adSlot, {}, false),
 			vastXml: vastResponse?.xml,
 		});
-	}
-
-	private setupOptimizelyExperiment() {
-		this.optimizely.addVariantToTargeting(
-			OPTIMIZELY_STRATEGY_RULES_EXPERIMENT,
-			OPTIMIZELY_STRATEGY_RULES_EXPERIMENT_VARIANTS.UNDEFINED,
-		);
-
-		const variant = this.optimizely.getVariant(OPTIMIZELY_STRATEGY_RULES_EXPERIMENT);
-
-		if (!variant) {
-			return;
-		}
-
-		this.optimizely.addVariantToTargeting(OPTIMIZELY_STRATEGY_RULES_EXPERIMENT, variant);
-		context.set(
-			'options.video.enableStrategyRules',
-			variant === OPTIMIZELY_STRATEGY_RULES_EXPERIMENT_VARIANTS.ENABLED,
-		);
 	}
 }
